`tools/importer-rest-api-specs`: moving the Terraform Helpers package into the Terraform package
