+++ tools/importer-rest-api-specs/components/terraform/helpers/snake_case.go
